+++ CHANGELOG.md
@@ -12,6 +12,7 @@ HumHub Changelog
 - Fix #8056: Fix linked badge icon
 - Fix #8059: Prevent previous message reused on attachment replies
 - Fix #8060: Double HTML encoding of the Tour widget menu "remove" item
+- Fix #8062: Fix global topics
 
 1.18.1 (March 2, 2026)
 ----------------------